chore: add the test to a github action.
Adding a matrix of supported operating systems and python versions
should give some confidence that the package is working as expected.

I have splitted the build / install / test targets to be able to do
the testing without having extra build dependencies (especially
patchelf which auditwheel depends on and is not a python dependency).

I'd like to also add some formatting and linting automation next.